refactor(Instantiator): move some methods to traits
The body of the instantiator object weren't probably the best place to
have the logic for creating the types tree or to map that tree to
a values tree.

I have moved them to traits that are used in the instantiator object, is
a cleaner implementation.
