+++ src/test/scala/com/pepegar/instantiator/utils/common.scala
@@ -1,0 +1,6 @@
+package com.pepegar.instantiator.utils
+
+case class A(b: B, c: C)
+case class B(d: D)
+case class C(d: D)
+case class D(left: Int, right: Float)
